Move types to own files
