+++ ModularPipelines/Engine/IDependencyChainProvider.cs
@@ -1,0 +1,8 @@
+﻿using ModularPipelines.Models;
+
+namespace ModularPipelines.Engine;
+
+internal interface IDependencyChainProvider
+{
+    IReadOnlyList<ModuleDependencyModel> ModuleDependencyModels { get; }
+}
