+++ ModularPipelines/Engine/IDependencyPrinter.cs
@@ -1,0 +1,6 @@
+﻿namespace ModularPipelines.Engine;
+
+internal interface IDependencyPrinter
+{
+    void Print();
+}
